fixed di problems

+++ shared/src/commonMain/kotlin/com/ancraz/mywallet_mult/data/database/dao/TransactionDao.kt
@@ -24,10 +24,10 @@ interface TransactionDao {
 
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun updateTransaction(transaction: TransactionEntity): Long
+    suspend fun updateTransaction(transaction: TransactionEntity): Int
 
 
     @Delete
-    suspend fun deleteTransaction(transaction: TransactionEntity): Long
+    suspend fun deleteTransaction(transaction: TransactionEntity): Int
 
 }
